refactor(shared-react): consolidate structure-protection baseline tests
Move the "disabled / not protected" cases out of the keyboard/non-keydown
suites and into the baseline test under a "plugin disabled — behaves like
baseline" describe, so unprotected behavior lives in one place. The keyboard
suite's testEnvironment now always renders the plugin as protected.

Also alphabetize StructureProtectionPlugin within Editor.tsx's plugin block
(no behavioral impact — its only same-priority co-registrant handles a
disjoint key set) and drop a stale spec reference from a comment.
