+++ libs/shared-react/src/plugins/usj/StructureProtectionPlugin.test.tsx
@@ -27,7 +27,7 @@ import { $createParaNode, ParaNode } from "shared";
 describe("StructureProtectionPlugin — keyboard", () => {
   it("blocks Backspace-at-start merge when protected", async () => {
     let t2: TextNode;
-    const { editor } = await testEnvironment(true, () => {
+    const { editor } = await testEnvironment(() => {
       t2 = $createTextNode("second");
       $getRoot().append(
         $createParaNode("p").append($createTextNode("first")),
@@ -43,27 +43,9 @@ describe("StructureProtectionPlugin — keyboard", () => {
     });
   });
 
-  it("allows Backspace-at-start merge when NOT protected", async () => {
-    let t2: TextNode;
-    const { editor } = await testEnvironment(false, () => {
-      t2 = $createTextNode("second");
-      $getRoot().append(
-        $createParaNode("p").append($createTextNode("first")),
-        $createParaNode("q").append(t2),
-      );
-    });
-    updateSelection(editor, t2!, 0);
-
-    await pressKey(editor, "Backspace", 0);
-
-    editor.getEditorState().read(() => {
-      expect($getRoot().getChildrenSize()).toBe(1); // merged
-    });
-  });
-
   it("blocks Enter (paragraph split) when protected", async () => {
     let t1: TextNode;
-    const { editor } = await testEnvironment(true, () => {
+    const { editor } = await testEnvironment(() => {
       t1 = $createTextNode("abcdef");
       $getRoot().append($createParaNode("p").append(t1));
     });
@@ -81,7 +63,7 @@ describe("StructureProtectionPlugin — non-keydown vectors", () => {
   it("blocks controlled text insertion over a selection spanning a block boundary when protected", async () => {
     let t1: TextNode;
     let t2: TextNode;
-    const { editor } = await testEnvironment(true, () => {
+    const { editor } = await testEnvironment(() => {
       t1 = $createTextNode("first");
       t2 = $createTextNode("second");
       $getRoot().append($createParaNode("p").append(t1), $createParaNode("q").append(t2));
@@ -97,29 +79,10 @@ describe("StructureProtectionPlugin — non-keydown vectors", () => {
     });
   });
 
-  it("allows controlled text insertion spanning a boundary when NOT protected", async () => {
-    let t1: TextNode;
-    let t2: TextNode;
-    const { editor } = await testEnvironment(false, () => {
-      t1 = $createTextNode("first");
-      t2 = $createTextNode("second");
-      $getRoot().append($createParaNode("p").append(t1), $createParaNode("q").append(t2));
-    });
-    updateSelection(editor, t1!, 0, t2!, 6);
-
-    await act(async () => {
-      editor.dispatchCommand(CONTROLLED_TEXT_INSERTION_COMMAND, "x");
-    });
-
-    editor.getEditorState().read(() => {
-      expect($getRoot().getChildrenSize()).toBe(1); // replaced/merged
-    });
-  });
-
   it("blocks insertion over a selection containing a verse marker when protected", async () => {
     let para: ParaNode;
     let t1: TextNode;
-    const { editor } = await testEnvironment(true, () => {
+    const { editor } = await testEnvironment(() => {
       para = $createParaNode("p");
       t1 = $createTextNode("text");
       $getRoot().append(para.append($createImmutableVerseNode("1"), t1));
@@ -143,7 +106,7 @@ describe("StructureProtectionPlugin — non-keydown vectors", () => {
   it("consumes CUT over an unsafe selection when protected (low-priority spy not reached)", async () => {
     let t1: TextNode;
     let t2: TextNode;
-    const { editor } = await testEnvironment(true, () => {
+    const { editor } = await testEnvironment(() => {
       t1 = $createTextNode("first");
       t2 = $createTextNode("second");
       $getRoot().append($createParaNode("p").append(t1), $createParaNode("q").append(t2));
@@ -158,32 +121,12 @@ describe("StructureProtectionPlugin — non-keydown vectors", () => {
     unregister();
 
     expect(spy).not.toHaveBeenCalled(); // HIGH handler blocked propagation
-  });
-
-  it("does NOT consume CUT when not protected (low-priority spy reached)", async () => {
-    let t1: TextNode;
-    let t2: TextNode;
-    const { editor } = await testEnvironment(false, () => {
-      t1 = $createTextNode("first");
-      t2 = $createTextNode("second");
-      $getRoot().append($createParaNode("p").append(t1), $createParaNode("q").append(t2));
-    });
-    updateSelection(editor, t1!, 0, t2!, 6);
-
-    const spy = vi.fn(() => true); // claim handled so RichText's own cut does nothing
-    const unregister = editor.registerCommand(CUT_COMMAND, spy, COMMAND_PRIORITY_LOW);
-    await act(async () => {
-      editor.dispatchCommand(CUT_COMMAND, null);
-    });
-    unregister();
-
-    expect(spy).toHaveBeenCalled(); // plugin inert; propagation continued
   });
 
   it("consumes DRAGSTART over an unsafe selection when protected (low-priority spy not reached)", async () => {
     let t1: TextNode;
     let t2: TextNode;
-    const { editor } = await testEnvironment(true, () => {
+    const { editor } = await testEnvironment(() => {
       t1 = $createTextNode("first");
       t2 = $createTextNode("second");
       $getRoot().append($createParaNode("p").append(t1), $createParaNode("q").append(t2));
@@ -199,31 +142,11 @@ describe("StructureProtectionPlugin — non-keydown vectors", () => {
 
     expect(spy).not.toHaveBeenCalled(); // HIGH handler blocked propagation
   });
-
-  it("does NOT consume DRAGSTART when not protected (low-priority spy reached)", async () => {
-    let t1: TextNode;
-    let t2: TextNode;
-    const { editor } = await testEnvironment(false, () => {
-      t1 = $createTextNode("first");
-      t2 = $createTextNode("second");
-      $getRoot().append($createParaNode("p").append(t1), $createParaNode("q").append(t2));
-    });
-    updateSelection(editor, t1!, 0, t2!, 6);
-
-    const spy = vi.fn(() => true);
-    const unregister = editor.registerCommand(DRAGSTART_COMMAND, spy, COMMAND_PRIORITY_LOW);
-    await act(async () => {
-      editor.dispatchCommand(DRAGSTART_COMMAND, null as unknown as DragEvent);
-    });
-    unregister();
-
-    expect(spy).toHaveBeenCalled(); // plugin inert; propagation continued
-  });
 });
 
-async function testEnvironment(isStructureProtected: boolean, $initialEditorState: () => void) {
+async function testEnvironment($initialEditorState: () => void) {
   return baseTestEnvironment(
     $initialEditorState,
-    <StructureProtectionPlugin isStructureProtected={isStructureProtected} />,
+    <StructureProtectionPlugin isStructureProtected={true} />,
   );
 }
